feat: Add QR code generator functionality
This commit adds the functionality to generate QR codes for the given input text. It includes a new file `main.py` which contains the code for generating QR codes using the `qrcode` library. The user can input the data to be encoded and the filename to save the QR code as. The generated QR code is saved as an image file in the same directory. Additionally, a `requirements.txt` file is included to specify the dependency on the `qrcode` library.

The purpose of this commit is to provide a simple QR code generator that can be easily used by users to generate QR codes for various purposes.